feat: add Linux builds
Massive thanks to Yosh for figuring out the build part and opening the
PR #2. This made making this build process super trivial.
Surprisingly, some dependencies weren't present on the default ubuntu
image for the runner, which is weird given that image is as bloated as
it can get.
In any case, this builds really fast and as expected. The main issue is
getting ARM64 builds to work as some march values are hardcoded in some
places and I don't feel like fighting with it tonight. ARM64 is probably
way out of scope for the "making stuff work" phase anyways.